SDL2: Make word jumping/deletion work in console

+++ src/helpers/utf8.h
@@ -44,8 +44,8 @@ public:
                                  unsigned int i_numChar);
   static unsigned int utf8_length(const std::string &i_a);
   static std::string utf8_substring(const std::string &i_a,
-                                    unsigned int i_numChar,
-                                    unsigned int i_nbChars);
+                                    size_t i_numChar,
+                                    size_t i_nbChars);
   /* substring with absolute indexes instead of a length */
   static std::string utf8_substring_abs(const std::string &str,
                                     size_t start,
